# Concurrent cache of generated classes