--- a/cglib/src/main/java/net/sf/cglib/core/ClassesKey.java
+++ b/cglib/src/main/java/net/sf/cglib/core/ClassesKey.java
@@ -16,7 +16,7 @@
 package net.sf.cglib.core;
 
 public class ClassesKey {
-    private static final Key FACTORY = (Key)KeyFactory.create(Key.class, KeyFactory.OBJECT_BY_CLASS);
+    private static final Key FACTORY = (Key)KeyFactory.create(Key.class);
     
     interface Key {
         Object newInstance(Object[] array);
@@ -26,6 +26,21 @@ public class ClassesKey {
     }
 
     public static Object create(Object[] array) {
-        return FACTORY.newInstance(array);
+        return FACTORY.newInstance(classNames(array));
+    }
+
+    private static String[] classNames(Object[] objects) {
+        if (objects == null) {
+            return null;
+        }
+        String[] classNames = new String[objects.length];
+        for (int i = 0; i < objects.length; i++) {
+            Object object = objects[i];
+            if (object != null) {
+                Class<?> aClass = object.getClass();
+                classNames[i] = aClass == null ? null : aClass.getName();
+            }
+        }
+        return classNames;
     }
 }