--- a/cglib/src/main/java/net/sf/cglib/core/Customizer.java
+++ b/cglib/src/main/java/net/sf/cglib/core/Customizer.java
@@ -17,6 +17,12 @@ package net.sf.cglib.core;
 
 import org.objectweb.asm.Type;
 
-public interface Customizer {
+/**
+ * Customizes key types for {@link KeyFactory} when building equals, hashCode, and toString.
+ * For customization of field types, use {@link FieldTypeCustomizer}
+ *
+ * @see KeyFactory#CLASS_BY_NAME
+ */
+public interface Customizer extends KeyFactoryCustomizer {
     void customize(CodeEmitter e, Type type);
 }