--- a/cglib/src/main/java/net/sf/cglib/core/HashCodeCustomizer.java
+++ b/cglib/src/main/java/net/sf/cglib/core/HashCodeCustomizer.java
@@ -1,0 +1,12 @@
+package net.sf.cglib.core;
+
+import org.objectweb.asm.Type;
+
+public interface HashCodeCustomizer extends KeyFactoryCustomizer {
+    /**
+     * Customizes calculation of hashcode
+     * @param e code emitter
+     * @param type parameter type
+     */
+    boolean customize(CodeEmitter e, Type type);
+}