--- a/cglib/src/main/java/net/sf/cglib/core/KeyFactoryCustomizer.java
+++ b/cglib/src/main/java/net/sf/cglib/core/KeyFactoryCustomizer.java
@@ -1,0 +1,7 @@
+package net.sf.cglib.core;
+
+/**
+ * Marker interface for customizers of {@link KeyFactory}
+ */
+public interface KeyFactoryCustomizer {
+}