--- a/cglib/src/main/java/net/sf/cglib/core/internal/Function.java
+++ b/cglib/src/main/java/net/sf/cglib/core/internal/Function.java
@@ -1,0 +1,5 @@
+package net.sf.cglib.core.internal;
+
+public interface Function<K, V> {
+    V apply(K key);
+}