… cglib/src/main/java/net/sf/cglib/core/internal/LoadingCache.java
… cglib/src/main/java/net/sf/cglib/proxy/CallbackInfo.java
… cglib/src/test/java/net/sf/cglib/beans/TestBeanMap.java
… cglib/src/test/java/net/sf/cglib/proxy/TestEnhancer.java
… cglib/src/test/java/net/sf/cglib/proxy/TestMixin.java
… cglib/src/test/java/net/sf/cglib/reflect/TestFastClass.java
… pom.xml