This fix removes `synchronization` in case relevant class is already cached.

The change comes on top of "fix permgen" (it reuses `WeakIdentityKey`)

Measurements (java 1.8u60, i7-4960HQ CPU @ 2.60GHz):

# 1 Thread

Response time is 15 times better.
Heap allocation is 5 times less.

Before:
```
Benchmark                                       Mode  Cnt     Score     Error   Units
BeansBenchmark.newInstance                      avgt    5  2135,987 ± 262,118   ns/op
BeansBenchmark.newInstance:·gc.alloc.rate.norm  avgt    5  1240,021 ±   0,168    B/op

```

After:
```
Benchmark                                       Mode  Cnt     Score     Error   Units
BeansBenchmark.newInstance                      avgt    5   142,438 ±   8,209   ns/op
BeansBenchmark.newInstance:·gc.alloc.rate.norm  avgt    5   256,001 ±   0,012    B/op
```

# Multiple threads

Good scalability in range of 1-2-4 threads.
Degradation after 4 is somewhat expected as the machine has just 4 hw cores.

Before:
```
Benchmark                   Mode Thr Cnt      Score      Error   Unit
BeansBenchmark.newInstance  avgt   1   5   2518,232 ±  148,755  ns/op
BeansBenchmark.newInstance  avgt   2   5   3207,201 ±  231,017  ns/op
BeansBenchmark.newInstance  avgt   4   5   7842,192 ±  528,633  ns/op
BeansBenchmark.newInstance  avgt   8   5  16380,357 ±  941,975  ns/op
```

After:
```
Benchmark                   Mode Thr Cnt      Score      Error   Unit
BeansBenchmark.newInstance  avgt   1   5    143,228 ±   15,594  ns/op
BeansBenchmark.newInstance  avgt   2   5    161,668 ±    2,057  ns/op
BeansBenchmark.newInstance  avgt   4   5    152,933 ±    2,753  ns/op
BeansBenchmark.newInstance  avgt   6   5    242,031 ±   42,862  ns/op
BeansBenchmark.newInstance  avgt   8   5    298,683 ±   39,506  ns/op
```