@vlsi 2015-11-16
Please, review the PR & approach.

In a couple of words:
1) I made class generation cache shared across all the `Source` instances. I think it simplifies code.
2) When new classLoader is added to the list, then `copy-on-write under global lock WeakHashMap` is used. I believe it should be very rare. `WeakHashMap` allows to have weak reference to the `ClassLoader` while avoiding creation of `WeakReference(classLoader)` for `map.get(...)` purpose.
3) Inside the classloader, `ConcurrentHashMap` is used, so readers do not block each other (so this is the major improvement over previous plain `HashMap`)
4) For repeatable object creation, separate `Factory` class is created and reused behind the scenes. Regular "proxy" is always created first, then factory is created if needed.

Even though the fix passes all the tests, it might make sense adding explicit API to get `Factory` implementation out of `Enhancer`. For instance `Enhancer#createFactory()` like the current `#createClass()`.