@raphw 2016-02-05
The cache is definetly a pain-point in cglib. This is a real improvement! Request looks good!